Edit Character dispatches per faction: chassis, the fold, and the two slots a create dress has no room for
Part of #2537.

`editCharacter` joins `FactionManifest` and `SURFACE_KEYS`, na registers the
folded responsive `DefaultEditCharacter`, and `EditCharacter.tsx` becomes a
dispatcher on the EDITED character's slug. The mobile-only copy is gone;
`mobileEditCharacter` stays retired.

The faction row and the destructive action live in `editCharacterSlots.tsx`,
designed once so the seven-faction fan-out inherits them.
