+++ frontend/src/pages/characterPaths/__tests__/disabledControlContrast.test.ts
@@ -81,7 +81,7 @@ const SITES = [
   'archetypes/SnideCreateCharacter.tsx',
   'archetypes/UaCreateCharacter.tsx',
   'archetypes/WowCreateCharacter.tsx',
-  'mobileArchetypes/DefaultEditCharacter.tsx',
+  'archetypes/DefaultEditCharacter.tsx',
 ]
 
 function source(file: string): string {
